requirements.txt
# Minimal requirements for Streamlit Cloud
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
chromadb>=0.4.15
pinecone-client>=3.0.0
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
plotly>=5.15.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.4.0
